backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
python-jose[cryptography]
passlib[argon2]
pydantic
pydantic-settings
sqladmin
google-generativeai
requests
PyMuPDF
PyPDF2
python-dotenv
gunicorn
jinja2
itsdangerous
WTForms
markupsafe
starlette
typing-extensions
annotated-types
certifi
charset-normalizer
idna
urllib3
fastapi-mail
authlib
httpx